Move ComparisonPassed() methods from WCSimRootGeom.cc and WCSimRootEvent.cc to new files WCSimRootTools.cc and WCSimRootTools.hh
